Fix some of admin-handlers-users and add some errors
