+++ cmd/typed-errors.go
@@ -69,6 +69,10 @@ var errNoSuchUser = errors.New("Specified user does not exist")
 // to use a LDAP command.
 var errNoSuchUserLDAPWarn = errors.New("Specified user does not exist. If you meant a user in LDAP please use command under `mc idp ldap`")
 
+// error returned by IAM when a builtin IDP command is used when they could mean
+// to use an LDAP command for groups.
+var errNoSuchGroupLDAPWarn = errors.New("Specified group does not exist. If you meant a group in LDAP please use command under `mc idp ldap`")
+
 // error returned when service account is not found
 var errNoSuchServiceAccount = errors.New("Specified service account does not exist")
